backend/requirements.txt
astroid==2.4.2
Click==7.1.2
ecdsa==0.16.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.12.0
SQLAlchemy==1.3.22
typed-ast==1.4.2
Werkzeug==1.0.1
wrapt==1.12.1
Flask-Cors==3.0.9